RavenDB-13110 - Incoming Replication Handler - initial abstraction
